+++ app/controllers/admin/films_controller.rb
@@ -10,7 +10,7 @@ class Admin
       if params[:edition_id].present?
         @edition = Edition.friendly.find(params[:edition_id])
         @festival = @edition.festival
-        @selections = @edition.selections.includes(:category, :film).where(film: @films).order("films.title")
+        @selections = @edition.selections.includes(:category, :film).where(film: @films).order("films.sort_title")
 
         render(
           partial: "admin/selections/search_results",